+++ src/__playground__/worker.ts
@@ -1,9 +1,7 @@
-import { createWriteStream } from 'fs';
-import { ParentThread, define } from 'nanolith';
+import { define } from 'nanolith';
 
 export const worker = await define({
-    async wait() {
-        await new Promise((r) => setTimeout(r, 5e3));
-        console.log('done');
+    async throw() {
+        throw new Error();
     },
 });