+++ src/constants/messages.ts
@@ -4,21 +4,20 @@
  */
 export const enum ParentThreadMessageType {
     /**
-     * To be used when sending a message to a worker from
-     * the main thread.
+     * When sending a message to a worker from the main thread.
      */
     Message,
     /**
-     * To be used when calling a task in a service worker.
+     * When calling a task in a service worker.
      */
     Call,
     /**
-     * To be used when posting a message to a service worker
+     * When posting a message to a service worker
      * notifying it to exit its process immediately.
      */
     Terminate,
     /**
-     * To be used when passing `Messenger` objects to workers.
+     * When passing `Messenger` objects to workers.
      */
     MessengerTransfer,
 }
@@ -29,48 +28,47 @@ export const enum ParentThreadMessageType {
  */
 export const enum WorkerMessageType {
     /**
-     * To be used when sending a message from a worker to
+     * When sending a message from a worker to
      * the parent thread.
      */
     Message,
     /**
-     * To be used when a task worker has returned a
+     * When a task worker has returned a
      * value and posting it back to the parent thread.
      */
     TaskReturn,
     /**
-     * To be used when a task worker has failed and
+     * When a task worker has failed and
      * an error has been thrown.
      */
     TaskError,
     /**
-     * To be used when a called task in a service worker has
+     * When a called task in a service worker has
      * returned a value and posting it back to the parent thread.
      */
     CallReturn,
     /**
-     * To be used when a called task in a service worker has
+     * When a called task in a service worker has
      * failed and posting the error back to the parent thread.
      */
     CallError,
     /**
-     * To be used when notifying the parent thread that a `Messenger`
+     * When notifying the parent thread that a `Messenger`
      * object has successfully been sent.
      */
     MessengerTransferSuccess,
     /**
-     * To be used when a service worker throws an exception and the parent
+     * When a service worker throws an exception and the parent
      * thread must be notified about it.
      */
     WorkerException,
     /**
-     * To be used to notify the parent thread that a service worker has
+     * To notify the parent thread that a service worker has
      * completed initialization and that it's ready to go.
      */
     Initialized,
     /**
-     * To be used for sending messages to the main thread notifying that a
-     * service has exited.
+     * For sending messages to the main thread notifying that a service has exited.
      */
     Exit,
 }