+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Improved performance for `ServiceCluster` methods `closeAll()`, `closeAllIdle()`, `use()`, and `activeServiceCalls()`.
 - `Service.close()` not doing early cleanup for callbacks.
 - `Messenger.close()` not doing early cleanup for stream event callbacks.
+- `define()` waiting to run the worker before resolving with the Nanolith API object.
 
 ### Added