+++ CHANGELOG.md
@@ -4,7 +4,31 @@ All notable changes to Nanolith since version 0.1.0 will be documented in this f
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+> Dates are in YYYY-D-M format.
+
 ## [Unreleased]
+
+### Fixed
+
+- Not calling `getDefaultPoolConcurrency` in README example.
+- Improved performance for `ServiceCluster` methods `closeAll()`, `closeAllIdle()`, `use()`, and `activeServiceCalls()`.
+- `Service.close()` not doing early cleanup for callbacks.
+- `Messenger.close()` not doing early cleanup for stream event callbacks.
+- `define()` waiting to run the worker before resolving with the Nanolith API object.
+- Missing `workerData.identifier` property in `PoolItemOptions` type.
+- `Service.call()` not cleaning up callbacks.
+- `Service.call()` not decreasing active count upon promise rejection.
+- Minor performance improvement for `Messenger` with streams.
+
+### Added
+
+- New tests for `ServiceCluster.use()`.
+
+### Removed
+
+- `ParentThreadTerminateMessageBody` type eliminated.
+
+## [0.4.4] - 2023-30-3
 
 ### Changed